Allows us to not save problem and solver data.
- Does not change default behavior.
- If we set
  - e["Save"]["Solver"] = False
  - e["Save"]["Problem"] = False
only the first and last generation of the problem and solver
- configuration will be saved;
In order to ensure that we have the first generation with info for
plotting scripts and last in order to reload experiments for further runs.
Why? If we have really big data sets and we want to store for example
- training and validation losses for each generation automatically in
- e["Results"] as be Korali intended, we will also store the problem and
- solver for each generation which may lead to 1000s of files of multiple
- 100Mb or even Gb.