Drops termination criteria once they are use.
terminationCriteria is an vector that only holds the string literal of
the termination it is not responsible for termination and will be pushed
back into the _terminationCriteria string once the termination criteria
is met again.
As such it should be removed once it is printed to the screen.
It should not change the default behavior but leads to issues otherwise.