Adds KORALI experiment profiler.
- adds macros to add profling timestamps.
- adds also profile to process the results files.

Can be used to profile any module that:
 1. has access to an experiment:
 2. includes the experiment header somehow

KORALI_START_PROFILE("name of profile", _k);
    .. comde code to run
KORALI_STOP_PROFILE("name of profile", _k);

Will be stored in the experiment in a vector for each generation.
Each entry is an unorderd map with the profile name and a vector of
length function calls in this generation.
Where each of this elements is a pair with.
 - first: start time relative to program start
 - second: duration.

Can be processed using:
python -m korali.profiler --input "...../_korali_result/latest" --profile-experiment [--plot]

TODO: move into engine profile instead of inside of experiment results.